requirements.txt
# Core LLM
openai>=1.0.0

# Embeddings & Vector Store
fastembed>=0.2.0
faiss-cpu>=1.7.4

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0

# Data Validation & Config
pydantic>=2.0.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
httpx>=0.26.0
